Create restore.ps1
This is a powershell script written for school. It was used on server in a lab environment running Active Directory services and SQL Server. Unfortunately it cannot be tested since it required that lab environment. It creates new objects in active directory. It also creates a new database in SQL and new table in the new database; Then it inserts data into the new table.